Replaced reinterpret_cast with static_cast in libprocess.
Casting from a virtual base class to a child class should not be
done with reinterpret_cast. We could use dynamic_cast as well, but
that has a performance cost; in this case, we know exactly which
child class the pointer points to, so the performance cost and
additional safety offered by dynamic_cast is not necessary.

Review: https://reviews.apache.org/r/45489