Updated release/versioning docs for "-dev" version scheme.
Mesos now uses the "-dev" prerelease label to identify code in the
master branch in the period before the next release branch has been
cut.

For example, after creating the 1.3.x release branch, we previously
updated the version number in the master branch to be "1.4.0". The new
policy is to instead use the version number "1.4.0-dev" in the master
branch. When the release branch for 1.4.x is created, the version number
will then be updated to remove the "-dev" prerelease label.

This commit also makes a few minor procedural changes to the release
process -- for example, we now create the release branch first and then
tag a commit in that branch, rather than first tagging a release and
then creating the branch from the tag.

Review: https://reviews.apache.org/r/59068