Prevented reserve/create with hierarchical roles on a non-capable agent.
This patch makes it such that reserve / create operations involving
resources with a hierarchical role are invalid if they are attempted on
an agent without a HIERARCHICAL_ROLE capability. Such attempts from an
operator endpoint will result in a \`BadRequest\` response. On the
frameworks side, we've opted to not offer resources from
a non-HIERARCHICAL_ROLE agent to hierarchical roles (ca8d33ab).

If we were to undo ca8d33ab, an attempt from a framework to perform such
an operation would be considered invalid due to this patch and the
operation would be dropped by the master.

Review: https://reviews.apache.org/r/60062/