Added a hack to avoid printing out the `Resource.role` field.
With reservation refinement, we use the \`Resource.reservations\` field
to express the reservation state. Due to the fact that \`Resource.role\`
is an optional field with a default value, our generic Protobuf to
JSON converter prints out the \`role\` field even if it's not set.
In order to mitigate the confusion for the users, we introduce a hack
for now to conditionally remove the \`role\` field.

Review: https://reviews.apache.org/r/60036/