Resources: Updated the validation logic.
This patch updates \`Resources::validate\`, and \`resource::validate\`.

The \`Resources::validate\` function ensures that the resources are
either in the 'pre-' or 'post-' reservation-refinement formats,
but not both.

The \`resource::validate\` function takes an optional framework
capabilities and ensures that the resources format matches
the framework's RESERVATION_REFINEMENT capbility.

Review: https://reviews.apache.org/r/60013/