Fixed subtle bug in GroupTest.ConnectTimer, GroupTest.TimerCleanup.
These tests looked for dispatches to `GroupProces::expired` as a way to
determine when the current Zk session has expired. However, the previous
implementation of the connect timer (`GroupProcess::timedout`) invoked
`GroupProcess::expired` directly, which meant that an `EXPECT_DISPATCH`
on the `expired` method should not fire.

However, a separate bug in the implementation of `EXPECT_DISPATCH` meant
that the expectation was actually being satisfied by a dispatch to
`GroupProcess::timedout`, which meant the test happened to work (!).

Fix this by changing `GroupProcess::timedout` to dispatch to `expired`
rather than invoking it directly.